+++ spec/arachni/element/link_template/dom_spec.rb
@@ -60,14 +60,6 @@ describe Arachni::Element::LinkTemplate::DOM do
         end
     end
 
-    %w(encode decode).each do |m|
-        describe "##{m}" do
-            it 'returns the string as is' do
-                subject.send( m, 'blah' ).should == 'blah'
-            end
-        end
-    end
-
     describe '#parent' do
         it 'returns the parent element' do
             subject.parent.should be_kind_of Arachni::Element::LinkTemplate