+++ spec/support/servers/checks/active/unvalidated_redirect_dom.rb
@@ -99,7 +99,7 @@ get '/cookie/straight' do
                         var c = ca[i].trim();
 
                         if( c.indexOf( name ) == 0 ) {
-                            return c.substring( name.length, c.length )
+                            return decodeURIComponent( c.substring( name.length, c.length ) )
                         }
                     }